requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0
requests>=2.28.0
beautifulsoup4>=4.12.0
streamlit>=1.28.0
plotly>=5.17.0
seaborn>=0.12.0
matplotlib>=3.7.0
jupyter>=1.0.0
python-dotenv>=1.0.0
lightgbm>=4.1.0
textblob>=0.17.0